Refactor : index.jsx 수정.
